Made TCache automatically grow the line size up to 16MB, based on incoming cache requests.  Initial cache layout is 1 line with 4KB line size.

+++ yamka/yamkaStdInt.cpp
@@ -382,7 +382,7 @@ namespace Yamka
       return 0;
     }
     
-    return storage.skip(numBytesToLoad + 1);
+    return (unsigned int)storage.skip(numBytesToLoad + 1);
   }
   
   //----------------------------------------------------------------
